adamantium/demo/cifar10_jit/requirements.txt
# --------- pytorch --------- #
torch>=2.0.0+cpu
torchvision>=0.15.0+cpu

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

# --------- deployment --------- #
gradio